% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCrosses.R
\name{buildCrosses}
\alias{buildCrosses}
\title{Create a data frame with all possible crosses}
\usage{
buildCrosses(moms, dads, keep, criterion, K)
}
\arguments{
\item{moms}{vector of possible moms to use}

\item{dads}{vector of possible dads to use}

\item{keep}{vector of genotypes to keep}

\item{criterion}{data frame with variable to maximize}

\item{K}{relationship matrix between all genotypes}
}
\value{
A data frame with all possible crosses.
}
\description{
Create a data frame based on given moms, dads, maximum relatedness, and maximum of individuals per relatedness cluster, return all crosses possible.
}
\examples{

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
