% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOCS.R
\name{calculateCoancestryRate}
\alias{calculateCoancestryRate}
\title{\code{calculateCoancestryRate} computes coancestry rate from current and future coancestries}
\usage{
calculateCoancestryRate(Actual, Future)
}
\arguments{
\item{Actual}{current coancestry}

\item{Future}{future coancestry}
}
\value{
coancestry rate
}
\description{
\code{calculateCoancestryRate} computes coancestry rate from current and future coancestries
}
