% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOCS.R
\name{Cont2Cross}
\alias{Cont2Cross}
\title{\code{Cont2Cross}}
\usage{
Cont2Cross(
  nContribution = NULL,
  AllowSelfing = FALSE,
  nCrosses = NULL,
  nProgeny = 1
)
}
\arguments{
\item{nContribution}{data frame with the id of the parent and the contribution of each one.}

\item{AllowSelfing}{Boolean. Is self allowed in the crosses? Default is FALSE.}

\item{nCrosses}{total number of crosses from the parents sets.}

\item{nProgeny}{number of progeny fro each cross generated.}
}
\value{
A list with the contribution of each parent and a valid mate plan.
}
\description{
Generates a list of crosses based on contribution values for a set of parents
}
