% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrib2Cross.R
\name{contrib2Cross}
\alias{contrib2Cross}
\title{Generates a mating plan based on parental contributions.}
\usage{
contrib2Cross(
  nContribution = NULL,
  AllowSelfing = FALSE,
  nCrosses = NULL,
  nProgeny = 1
)
}
\arguments{
\item{nContribution}{data frame with two columns: i. parent id and (ii) contribution of the parent.}

\item{AllowSelfing}{Boolean. Is self allowed in the crosses? Default is FALSE.}

\item{nCrosses}{total number of crosses from the parents sets.}

\item{nProgeny}{number of progeny for each cross generated.}
}
\value{
A mating plan based on the individuals and contributions of each one.
}
\description{
Generates a mating plan based on contribution values predetermined for a set of parents. In optimal contribution selection, a set of
parameters are used, such as inbreeding level, estimated breeding value, etc., into optimization algorithms to come up with a set of contributions of
each parent to the next generation. This function creates the mating plan based on the contribution of each parent to the next generation.
}
\examples{
\dontrun{
# 1. Loading the data
Contrib = data.frame(id = paste0('Par_',sample(1:20, 20)),
                     Contrib = sample(c(0:5), 20, replace = T))

# 2. Contributions
MatingPlan <- contrib2Cross(nContribution=Contrib,
                            AllowSelfing = FALSE,
                            nCrosses = 20,
                            nProgeny = 1)

# 3. Output
head(MatingPlan, 10)
}

}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
