% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOCS.R
\name{CRate.target}
\alias{CRate.target}
\title{\code{CRate.target} converts frontier degree to MaxCriterionPct}
\usage{
CRate.target(Degree = NULL, MaxObj.Coan, MinObj.Coan, CurrentCoancestry)
}
\arguments{
\item{Degree}{frontier degree ranging from zero to one hundred.}

\item{MaxObj.Coan}{maximum coancestry for the object.}

\item{MinObj.Coan}{minimum coancestry for the object.}

\item{CurrentCoancestry}{coancestry rate for the population.}
}
\value{
numeric, MaxCriterionPct (percentage of maximum criterion achieved (100 means we achieved the maximum possible selection criterion))
}
\description{
\code{CRate.target} converts frontier degree to MaxCriterionPct
}
