% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datGeneric_data.R
\docType{data}
\name{datGeneric}
\alias{datGeneric}
\title{Dataset generate from a generic population}
\format{
A data frame with 100 observations and 2 traits controlled by additive and dominance effects:
\describe{
\item{addEff}{Additive effects for each one of the markers.}
\item{domEff}{Dominance effects for each one of the markers.}
\item{Criterion}{BLUPs for each one of the 100 individuals present in the population, for both traits.}
\item{G}{A SNP-based relationship matrix for all individuals present in the population.}
\item{Map.In}{A data frame with the information regarding the SNP position in the genome. Three columns with: Chromosome, position, and marker id.}
\item{Markers}{A set of 1440 biallelic SNP markers coded in 0,1,2 for all individuals present in the population.}
\item{Phasedmarkers}{Phased haplotypes from the same set of individuals present in the population}
}
}
\source{
This dataset was simulated using AlphaSimR package (Gaynor et al. 2020). Two traits and 100 individuals with 12 chromosome pairs were generated.
}
\description{
Dataset generate from a generic population
}
\examples{
# Load the dataset
data("datGeneric")

# Additive effects
head(addEff)

# Dominance effects
head(domEff)

# Criterion
head(Criterion)

# Relationship Matrix
G[1:10,1:10]

# SNP position information
head(Map.In)

# Markers
Markers[1:10,1:10]

}
\references{
\emph{Gaynor, R. C., Gorjanc, G., & Hickey, J. M. (2021). AlphaSimR: an R package for breeding program simulations. G3, 11(2), jkaa017.}
}
