% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datLines_data.R
\docType{data}
\name{datLines}
\alias{datLines}
\title{Input example for SimpleMating package}
\format{
data.frame
}
\usage{
data(datLines)
}
\description{
Input example for SimpleMating package
}
\examples{
data(datLines)

}
\keyword{datasets}
