% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devolutePlan.R
\name{devolutePlan}
\alias{devolutePlan}
\title{Devolute mating plan to reduce relatedness}
\usage{
devolutePlan(
  data,
  K,
  n.cross = 200,
  max.cross = 4,
  min.cross = 2,
  max.cross.to.search = 1e+05,
  culling.pairwise.k = 0.25,
  n.threads = 1,
  n.steps = 200,
  min.target.k = NULL
)
}
\arguments{
\item{data}{data frame with possible crosses to filter out}

\item{K}{relationship matrix}

\item{n.cross}{number of crosses in the mating plan}

\item{max.cross}{maximum number of crosses per individual in the plan}

\item{min.cross}{minimum number of crosses per individual in the plan}

\item{max.cross.to.search}{maximum number of rows in the input to search for the solution}

\item{culling.pairwise.k}{don't allow crosses with more than this threshold of relatedness. Default is 0.25.}

\item{n.threads}{number of threads to run in parallel}

\item{n.steps}{number of steps in the evolution}

\item{min.target.k}{if target K is lower than this threshold, the algorithm stops.}
}
\value{
A list with a data frame with the steps and a data frame with the mating plan at the last iteration
}
\description{
Devolute mating plan to reduce relatedness
}
\examples{

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
