% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoluteCandidates.R
\name{evoluteCandidates}
\alias{evoluteCandidates}
\title{Pick candidates to reduce relatedness and maximize criterion}
\usage{
evoluteCandidates(
  criterion,
  K,
  n.candidates,
  step = c(5, 4, 3, 2, 1),
  n.steps = 200,
  n.try = 10000,
  n.burnIn = 10
)
}
\arguments{
\item{criterion}{data frame with variable to maximize}

\item{K}{relationship matrix}

\item{n.candidates}{number of crosses in the mating plan}

\item{n.try}{attempts to replace the current number of candidates (current step) to improve candidates list before moving for next phase}

\item{n.burnIn}{attempts to initialize the chain}

\item{steps}{numeric vector of steps in each evolution phase}
}
\value{
A list with final candidate list, and relationship and criterion evolution
}
\description{
Evolutionary algorithm to pick candidates for mating plan to reduce relatedness and maximize criterion
}
\examples{

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
