% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_simplemating-data.R
\docType{data}
\name{example_simplebreedingplan}
\alias{example_simplebreedingplan}
\title{Input example for SmallBreedingPlan package}
\format{
data.frame
}
\usage{
data(example_smallbreedingplan)
}
\description{
Input example for SmallBreedingPlan package
}
\examples{
data(example_simplebreedingplan)

}
\keyword{datasets}
