% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_GenMap.R
\docType{data}
\name{generic_GenMap}
\alias{generic_GenMap}
\title{generic_GenMap}
\format{
A data frame with 1440 rows and 3 variables.
\describe{
\item{\code{chr}}{Chromosome containing the locus}
\item{\code{pos}}{Genetic map position}
\item{\code{mkr}}{Unique identifier for locus}
}
}
\usage{
data(generic_GenMap)
}
\description{
A genetic map with the chromosome, position in the chromosome (centimorgan) and marker identification for the markers presented in the dataset. 12 chromosomes were simulated.
}
\details{
This genetic map is used in the prediction of a recombination map for the parental population.
}
\keyword{datasets}
