% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_geno.R
\docType{data}
\name{generic_geno}
\alias{generic_geno}
\title{generic_geno}
\format{
A data frame with 100 rows and 1440 variables.
}
\usage{
data(generic_geno)
}
\description{
Dataset containing the markers coded as 0,1,2 (aa, Aa, AA) for a population of 100 individuals genotyped for 1440 codominante SNPs markers for 12 chromosome pairs.
}
\details{
Data was simulated.
}
\keyword{datasets}
