% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_Phasedgeno.R
\docType{data}
\name{generic_Phasedgeno}
\alias{generic_Phasedgeno}
\title{generic_Phasedgeno}
\format{
A data frame with 200 rows and 1440 variables.
}
\usage{
data(generic_Phasedgeno)
}
\description{
Dataset with the phased state of a 1440 SNPs coded as 0 (reference allele) and 1 (alternate allele) across 12 chromosome pairs. The haplotypes are available for 100 individuals.
}
\details{
Dataset was simulated.
}
\keyword{datasets}
