% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsefAD.R, R/getUsefAD_mt.R
\name{getDiplotypes}
\alias{getDiplotypes}
\title{\code{getDiplotypes}
Function to estimates the diplotypes from a matrix of haplotypes parents}
\usage{
getDiplotypes(Markers)

getDiplotypes(Markers)
}
\arguments{
\item{Markers}{matrix containing the phased haplotypes for each parent in the parental population}
}
\description{
\code{getDiplotypes}
Function to estimates the diplotypes from a matrix of haplotypes parents

\code{getDiplotypes}
Function to estimates the diplotypes from a matrix of haplotypes parents
}
