% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getIndex}
\alias{getIndex}
\title{Estimates a selection index based on the genotypes values and weights}
\usage{
getIndex(Criterion, Weights = NULL, Scale = TRUE)
}
\arguments{
\item{Criterion}{matrix containing the BLUPs, estimated breeding value, or genetic value for all individuals and several traits.}

\item{Weights}{row vector containing the weights for each trait.}

\item{Scale}{logical. Default value is TRUE. If FALSE, it will not scale the traits.}
}
\value{
A index based on the traits and weights.
}
\description{
The function calculates the index-based trait using as input BLUPs, estimated breeding value, or genetic value for a set of individuals. Weights should be given.
}
\examples{
\dontrun{
# 1. Loading the data
data(generic_IndBLUP)

Crit <- as.matrix(generic_IndBLUP)

# 2. Index
trait_index <- getIndex(Criterion = Crit,
                        Weights = c(0.5, 0.5),
                        Scale = TRUE)

# 3. Output

head(trait_index, 10)
}

}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
