% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrossMPA.R
\name{getMPA}
\alias{getMPA}
\title{Create a data frame based on given mate plan, relatedness and mean parental average for each cross.}
\usage{
getMPA(MatePlan, Criterion, K = NULL, Weights = NULL, Scale = TRUE)
}
\arguments{
\item{MatePlan}{vector of possible crosses to make}

\item{Criterion}{data frame with variable to maximize}

\item{K}{relationship matrix between all individuals}

\item{Weights}{vector with the weights for each trait}

\item{Scale}{Boolean. If TRUE, the traits values will be scaled. Default is TRUE}
}
\value{
A data frame with all possible crosses.
}
\description{
Create a data frame based on given mate plan, relatedness and mean parental average for each cross.
}
\examples{
# 1.Loading the data
data('datGeneric')

# 2.Mating Plan
CrossPlan = planCross(TargetPop = colnames(G) )

# 3. Criterion
Crit = data.frame(Id = colnames(G),
                  Crit = Criterion[,1])

# 4.Single trait mean parental average
ST_mpa = getMPA(MatePlan = CrossPlan, 
                Criterion = Crit, 
                K=G)

head(ST_mpa, 20)

# 5. Criterion
CritMT = data.frame(Id = colnames(G),
                    Crit = Criterion)

# 6.Multi trait mean parental average
MT_mpa = getMPA(MatePlan = CrossPlan, 
                Criterion = CritMT, 
                K=G, 
                Scale = TRUE, 
                Weights = c(1,1))

head(MT_mpa, 20)

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com} & Marco A. Peixoto, \email{marco.peixotom@gmail.com}
}
