% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrossTGV.R
\name{getTGV}
\alias{getTGV}
\title{Prediction of total genetic value for a set of crosses}
\usage{
getTGV(
  MatePlan = NULL,
  Markers = NULL,
  addEff = NULL,
  domEff = NULL,
  Weights = NULL,
  Scale = TRUE,
  K = NULL
)
}
\arguments{
\item{MatePlan}{data frame with two columns indicating the crosses.}

\item{Markers}{matrix with markers information for all candidate parents,
coded as 0,1,2.}

\item{addEff}{vector with additive marker effects}

\item{domEff}{vector with dominance markers effects}

\item{Weights}{vector with the weights for each trait}

\item{Scale}{Boolean. If TRUE, the traits values will be scaled. Default is TRUE}

\item{K}{relationship matrix between all candidates to parent}
}
\value{
A data frame with total genetic value for each pair of
crosses presented in the MatePlan.
}
\description{
Predicts total genetic value for a set of crosses. In this case, we followed the formulae from Falconer and Mackay (1996).
}
\examples{

# 1. Loading the data
data('datGeneric')

# 2. Mating Plan
CrossPlan = planCross(colnames(G))

# 3. Single trait
ST_tgv = getTGV(MatePlan = CrossPlan, 
                Markers=Markers, 
                addEff=addEff[,1], 
                domEff=domEff[,1], 
                K = G)

head(ST_tgv, 20)

# 4. Multi trait
MT_tgv = getTGV(MatePlan = CrossPlan, 
                Markers=Markers, 
                addEff=addEff, 
                domEff=domEff, 
                Weights = c(0.2,0.8),
                K = G)

head(MT_tgv, 20)

}
\references{
\emph{Falconer, DS & Mackay TFC (1996). Introduction to quantitative genetics. Pearson Education India.}
}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
