% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsefA.R
\name{getUsefA}
\alias{getUsefA}
\title{Prediction of usefulness for a set of crosses (Single additive trait)}
\usage{
getUsefA(
  MatePlan,
  Markers,
  addEff,
  K,
  Map.In,
  linkDes = NULL,
  propSel = 0.05,
  Type = "DH",
  Generation = 1
)
}
\arguments{
\item{MatePlan}{data frame with the two columns indicating the crosses to estimates usefulness.}

\item{Markers}{matrix with markers information for all candidate parents,
coded as 0,1,2.}

\item{addEff}{column vector with additive marker effects.}

\item{K}{relationship matrix between all genotypes.}

\item{Map.In}{data frame with the genetic map information, i.e., Chromosome containing the locus, genetic map position, and unique identifier for locus.}

\item{linkDes}{Linkage disequilibrium matrix with the size of the total number of SNPs. This is optional, and it should be used only if the information on the genetic map is not available.}

\item{propSel}{Value representing the proportion of the selected individuals. Default is 0.05.}

\item{Type}{which kind of system of mating: "DH": doubled-haploids lines or
"RIL": Recombinant inbred lines.}

\item{Generation}{integer. Indicates the generation where the DH lines are generate or the RILs are extracted. According to Lehermeier et al. (2017)
DH derived from F1 generation is '1' and RILs from F2 generation is '1'. Also, for infinite generation a value superior to 10 should be used.}
}
\value{
A data frame with means, variances, and usefulness for each pair of
crosses presented in the MatePlan.
}
\description{
Predicts usefulness component for a set of crosses. It accounts for only one
trait controlled by additive effects. The variances were implemented according to Lehermeier et al. (2017). It accommodates Doubled-haploids (DH) and Recombinant
inbred lines (RILs) types. The genetic map is used to built a recombination map for the population (we implemented the Haldane map function internally).
}
\examples{
\dontrun{
# 1. Loading the dataset
data(lines_GenMap) # Genetic Map

data(lines_addEffects) # Additive effects

data(lines_Geno) # Markers

# 2. Parents
Parents <- rownames(lines_Geno)

# 3. Creating the mating plan
plan <- planCross(TargetPop = Parents,
                  MateDesign = "half")

# 4. Creating relationship matrix based on markers
ScaledMarkers <- scale(lines_Geno, scale = FALSE)

relMat <- (ScaledMarkers \%*\% t(ScaledMarkers)) / ncol(ScaledMarkers)

# 5. Calculating the usefulness of trait number 1
usef_add <- getUsefA(MatePlan = plan,
                     Markers = lines_Geno,
                     addEff = lines_addEffects[, 1],
                     Map.In = lines_GenMap,
                     K = relMat,
                     propSel = 0.05,
                     Type = "DH",
                     Generation = 1)

head(usef_add[[1]], 10)

head(usef_add[[2]], 10)
}

}
\references{
\emph{Lehermeier, C., de los Campos, Teyssèdre, S., & Schön, C. C. (2017). Genetic gain increases by applying the usefulness criterion with improved variance prediction in selection of crosses. Genetics, 207(4), 1651-1661.}

\emph{Bonk, S., Reichelt, M., Teuscher, F., Segelke, D., & Reinsch, N. (2016). Mendelian sampling covariability of marker effects and genetic values. Genetics Selection Evolution, 48(1), 1-11.}
}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
