% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsefAD.R
\name{getUsefAD}
\alias{getUsefAD}
\title{Predicts usefulness component for a set of crosses.}
\usage{
getUsefAD(
  MatePlan = NULL,
  Markers = NULL,
  addEff = NULL,
  domEff = NULL,
  Map.In = NULL,
  propSel = 0.05,
  Method = "Bonk"
)
}
\arguments{
\item{MatePlan}{data frame with the two columns indicating the crosses.}

\item{Markers}{matrix with markers information for all candidate parents, coded as 0,1,2. If Method is equal to Bonk or Wolfe, phased diplotypes should be given.}

\item{addEff}{data frame with additive marker effects for each trait (mxn, where 'm' is the number of individuals and 'n' represents the number of traits).}

\item{domEff}{data frame with dominance markers effects for each trait (mxn, where 'm' is the number of individuals and 'n' represents the number of traits).}

\item{Map.In}{data frame with the mapping information, i.e., chromosome, positioning and SNP id.}

\item{propSel}{Value representing the proportion of the selected individuals.}

\item{Method}{Which method should be used to calculates the progeny variances.}
}
\value{
A data frame with means, variances, and usefulness for each pair of
crosses presented in the MatePlan.
}
\description{
Prediction of usefulness for a set of cross using additive and dominance effects and a set of traits.
}
\examples{
# 1.Loading the dataset.
data('datGeneric')

# 2.Using just a subset for time purposes
Parents = colnames(G)[1:15]

# 3.Creating the mating plan
plan = planCross(TargetPop = Parents,
                 MateDesign = 'half')


# 4.Calculating the usefulness using Bonk method
usefBonk = getUsefAD(MatePlan = plan,
                     Markers = PhasedMarkers,
                     addEff = addEff[,1],
                     domEff = domEff[,1],
                     Map.In = Map.In,
                     propSel = 0.05,
                     Method = 'Bonk')

head(usefBonk, 10)

# 5.Calculating the usefulness using Wolfe method
usefWolfe = getUsefAD(MatePlan = plan,
                      Markers = PhasedMarkers,
                      addEff = addEff[,1],
                      domEff = domEff[,1],
                      Map.In = Map.In,
                      propSel = 0.05,
                      Method = 'Wolfe')

head(usefWolfe,10)

# 6.Calculating the usefulness using 'NonPhased' method
usefNonPhased = getUsefAD(MatePlan = plan,
                          Markers = Markers,
                          addEff = addEff[,1],
                          domEff = domEff[,1],
                          Map.In = Map.In,
                          propSel = 0.05,
                          Method = 'NonPhased')

head(usefNonPhased,10)

}
\references{
\emph{Lehermeier, C., de los Campos, Teyssèdre, S., & Schön, C. C. (2017). Genetic gain increases by applying the usefulness criterion with improved variance prediction in selection of crosses. Genetics, 207(4), 1651-1661.}

\emph{Bonk, S., Reichelt, M., Teuscher, F., Segelke, D., & Reinsch, N. (2016). Mendelian sampling covariability of marker effects and genetic values. Genetics Selection Evolution, 48(1), 1-11.}

\emph{Wolfe, M. D., Chan, A. W., Kulakow, P., Rabbi, I., & Jannink, J. L. (2021). Genomic mating in outbred species: predicting cross usefulness with additive and total genetic covariance matrices. Genetics, 219(3), iyab122.}
}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
