% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsefAD_mt.R
\name{getUsefAD_mt}
\alias{getUsefAD_mt}
\title{Prediction of usefulness for a set of crosses (Multi additive/dominance traits)}
\usage{
getUsefAD_mt(
  MatePlan,
  Markers,
  addEff,
  domEff,
  K,
  Map.In,
  linkDes = NULL,
  propSel = 0.05,
  Weights = NULL,
  Method = "Phased"
)
}
\arguments{
\item{MatePlan}{data frame with the two columns indicating the crosses.}

\item{Markers}{matrix with markers information for all candidate parents, coded as 0,1,2. If Method is equal 'Phased', phased haplotypes should be given and the markers is coded as 0 and 1.}

\item{addEff}{matrix with additive marker effects for each trait (mxn, where 'm' is the number of individuals and 'n' represents the number of traits).}

\item{domEff}{matrix with dominance markers effects for each trait (mxn, where 'm' is the number of individuals and 'n' represents the number of traits).}

\item{K}{relationship matrix.}

\item{Map.In}{data frame with the mapping information, i.e., Chromosome containing the locus, genetic map position, and unique identifier for locus.}

\item{linkDes}{Linkage disequilibrium matrix with the size of the total number of SNPs. This is optional, and it should be used only if the information on the genetic map is not available.}

\item{propSel}{Value representing the proportion of the selected individuals.}

\item{Weights}{Row vector containing the weights for each trait}

\item{Method}{Which method should be used to calculates the progeny variances.}
}
\value{
A data frame with means, variances, and usefulness for each pair of
crosses presented in the MatePlan.
}
\description{
Predicts usefulness component for a set of crosses. It accounts for only one
trait controlled by additive effects. The variances were implemented according to Lehermeier et al. (2017), Bonk et al. (2016), Wolfe et al. (2021), and Peixoto et al. (2024). The genetic map is used to
built a recombination map for the population (we implemented the Haldane map function). Two methods are implemented, one that uses phased haplotypes and another that uses non phased diplotypes.
Weights should be given.
}
\examples{
\dontrun{
# 1. Loading the dataset
data(generic_GenMap) # Genetic Map

data(generic_MrkEffects) # Additive effects

data(generic_Geno) # Markers

data(generic_Phasedgeno) # Phased haplotypes


# 2. Parents
Parents <- rownames(generic_Geno)

# 3.Creating the mating plan
plan <- planCross(TargetPop = Parents,
                  MateDesign = "half")

# 4. Creating relationship matrix based on markers
ScaledMarkers <- scale(generic_Geno, scale = FALSE)

relMat <- (ScaledMarkers \%*\% t(ScaledMarkers)) / ncol(ScaledMarkers)

# 5.Calculating the usefulness using 'Phased' method
usefPhased <- getUsefAD_mt(MatePlan = plan,
                           Markers = generic_Phasedgeno,
                           addEff = generic_MrkEffects[, 1:2],
                           domEff = generic_MrkEffects[, 3:4],
                           Map.In = generic_GenMap,
                           K = relMat,
                           propSel = 0.05,
                           Weights = c(0.2, 0.3),
                           Method = "Phased")

head(usefPhased[[1]] ,10)

head(usefPhased[[2]] ,10)

# 6.Calculating the usefulness using 'NonPhased' method
usefNonPhased <- getUsefAD_mt(MatePlan = plan,
                              Markers = generic_Geno,
                              addEff = generic_MrkEffects[, 1:2],
                              domEff = generic_MrkEffects[, 3:4],
                              Map.In = generic_GenMap,
                              K = relMat,
                              propSel = 0.05,
                              Weights = c(0.2, 0.3),
                              Method = "NonPhased")


head(usefNonPhased[[1]], 10)

head(usefNonPhased[[2]], 10)
}

}
\references{
\emph{Lehermeier, C., de los Campos, Teyssèdre, S., & Schön, C. C. (2017). Genetic gain increases by applying the usefulness criterion with improved variance prediction in selection of crosses. Genetics, 207(4), 1651-1661.}

\emph{Bonk, S., Reichelt, M., Teuscher, F., Segelke, D., & Reinsch, N. (2016). Mendelian sampling covariability of marker effects and genetic values. Genetics Selection Evolution, 48(1), 1-11.}

\emph{Wolfe, M. D., Chan, A. W., Kulakow, P., Rabbi, I., & Jannink, J. L. (2021). Genomic mating in outbred species: predicting cross usefulness with additive and total genetic covariance matrices. Genetics, 219(3), iyab122.}

\emph{Peixoto, Amadeu, Bhering, Ferrao, Munoz, & Resende Jr. (2024). SimpleMating:  R-package for prediction and optimization of breeding crosses using genomic selection}
}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
