% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOCS.R
\name{GOCS}
\alias{GOCS}
\title{Optimum Contribution Selection via quadratic programming.}
\usage{
GOCS(K, Criterion, nCross, Target = NULL, Degree = NULL, nProgeny = 1)
}
\arguments{
\item{K}{relationship matrix. It could be a SNP-based or pedigree-based
relationship matrix.}

\item{Criterion}{data frame with two columns: first column indicating the
individuals' name and the second with individuals' information taken as a
criterion to be optimized. Estimated breeding values, total genetic values,
BLUPs, or phenotypic records can be used.}

\item{nCross}{number of crosses to be estimated from a cohort of individuals.}

\item{Target}{target optimization. 'MaxCrit': Target to increase the selection
criterion. 'MinInb': target to decrease the inbreeding rates, or 'both': target
to increase the criterion while decrease the inbreeding rates.}

\item{Degree}{value for the degree which represents the angle for the line
in between the selection criterion increase and inbreeding rates decreasing.
Only valid for Target = 'both'.}

\item{nProgeny}{Size of the progeny that come from each cross.}
}
\value{
A valid Mating plan.
}
\description{
Function for selection of the best crosses via optimum contribution selection based on quadratic programming.
}
\examples{

# Loading the data
data('datGeneric')

# 1.Criterion
Crit = data.frame(Id = colnames(G),
                  BLUP = Criterion[,1])

# 2. Optimization targeting criterion
MatePlanCrit = GOCS(K = G, 
                    Criterion = Crit, 
                    nCross = 100, 
                    Target = 'MaxCrit')

# Individuals contribution
MatePlanCrit[[1]]

# Mating Plan
MatePlanCrit[[2]]

# Parents parameters
MatePlanCrit[[3]]

# New population parameters
MatePlanCrit[[4]]

# 3. Optimization targeting inbreeding
MatePlanInb = GOCS(K = G, 
                   Criterion = Crit, 
                   nCross = 100, 
                   Target = 'MinInb')

# Individuals contribution
MatePlanInb[[1]]

# Mating Plan
MatePlanInb[[2]]

# Parents parameters
MatePlanInb[[3]]

# New population parameters
MatePlanInb[[4]]

# 4. Optimization targeting both, criterion and inbreeding
MatePlan = GOCS(K = G, 
                Criterion = Crit, 
                nCross = 100, 
                Target = 'both',
                Degree = 80)

# Individuals contribution
MatePlan[[1]]

# Mating Plan
MatePlan[[2]]

# Parents parameters
MatePlan[[3]]

# New population parameters
MatePlan[[6]]

}
