% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsefA.R, R/getUsefA_mt.R
\name{Hal.Rec}
\alias{Hal.Rec}
\title{\code{Hal.Rec}
Function to calculate the recombination matrix from a genetic map based on Haldane (1909).}
\usage{
Hal.Rec(map)

Hal.Rec(map)
}
\arguments{
\item{map}{data frame with three columns: chromosome number, chromosome position, and markers number}
}
\description{
\code{Hal.Rec}
Function to calculate the recombination matrix from a genetic map based on Haldane (1909).

\code{Hal.Rec}
Function to calculate the recombination matrix from a genetic map based on Haldane (1909).
}
