% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_addEffects.R
\docType{data}
\name{lines_addEffects}
\alias{lines_addEffects}
\title{lines_addEffects}
\format{
A data frame with 1230 rows and 2 variables.
\describe{
\item{\code{add_trait1}}{Additive effects for trait one}
\item{\code{add_trait2}}{Additive effects for trait two}
}
}
\usage{
data(lines_addEffects)
}
\description{
Additive effects for 1230 markers and for two traits of interest. Each line represents one marker.
}
\details{
The SNP effects came from a Bayesian model implemented using both traits together. A total of 10 chromosomes were simulated.
}
\keyword{datasets}
