% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_geno.R
\docType{data}
\name{lines_geno}
\alias{lines_geno}
\title{lines_geno}
\format{
A data frame with 100 rows and 1230 variables.
}
\usage{
data(lines_geno)
}
\description{
Dataset containing the markers coded as 0,1,2 (aa, Aa, AA) for a population of doubled-haploid individuals genotyped for 1230 SNPs markers and 10 chromosomes.
}
\details{
Data was simulated.
}
\keyword{datasets}
