% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_IndBLUP.R
\docType{data}
\name{lines_IndBLUP}
\alias{lines_IndBLUP}
\title{lines_IndBLUP}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{\code{Genotypes}}{Gentoypes identification}
\item{\code{Trait1}}{BLUP for trait number 1}
\item{\code{Trait2}}{BLUP for trait number 2}
}
}
\usage{
data(lines_IndBLUP)
}
\description{
Best-linear unbiased prediction (BLUP) for 100 genotypes and two traits.
}
\details{
Data was simulated.
}
\keyword{datasets}
