% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Max_Avoid.R
\name{MaxAvoid}
\alias{MaxAvoid}
\title{Create a mating plan based on the maximum avoidance of inbreeding (Kimura and Crow, 1963)}
\usage{
MaxAvoid(nInd = NULL, nProgeny = 2L, Ind.ID = NULL)
}
\arguments{
\item{nInd}{Number of individuals to the crossed.}

\item{nProgeny}{Progeny number generated from each cross.}

\item{Ind.ID}{optional. Character vector with a list of candidate individuals. It will be
used to return the Mating plan with the genotype name.}
}
\value{
A data frame a valid Mating plan
}
\description{
In the maximum avoidance the individuals (parents) are arranged alternately
so that each individual is not mated with its neighbor in the right nor left.
}
\examples{
 
# 1.Loading the data
data('datGeneric')

# 2.Maximum avoidance
Plan = MaxAvoid(nInd=100, 
                nProgeny=1L, 
                Ind.ID=colnames(G))

# 3.Mating Plan
Plan

}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
