% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutatePlan.R
\name{mutatePlan}
\alias{mutatePlan}
\title{Evolute mating plan to reduce relatedness}
\usage{
mutatePlan(
  initialPlan,
  data,
  K,
  rate = 10,
  memory.mutation = 3,
  tries = 1000,
  perc.reduce.Y = 0,
  n.cross = 100,
  max.cross = 4,
  min.cross = 2,
  max.cross.to.search = 1e+05,
  culling.pairwise.k = 0.0675
)
}
\arguments{
\item{initialPlan}{initial plan to begin to mutate}

\item{data}{data frame with possible crosses to filter out}

\item{K}{relationship matrix}

\item{rate}{mutation rate (how many lines to substitute)}

\item{memory.mutation}{after how many successes to reset the mutation memory}

\item{tries}{maximum number of attempts to mutate}

\item{perc.reduce.Y}{fraction allowed to reduce Y during mutation}

\item{n.cross}{number of crosses in the mating plan}

\item{max.cross}{maximum number of crosses per individual in the plan}

\item{min.cross}{minimum number of crosses per individual in the plan}

\item{max.cross.to.search}{maximum number of rows in the input to search for the solution}

\item{culling.pairwise.k}{don't allow crosses with more than this threshold of relatedness. Default is 0.25.}
}
\value{
A list with a data frame with the steps and a data frame with the mating plan at the last iteration
}
\description{
Evolute mating plan to reduce relatedness
}
\examples{

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
