% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastThinning.R
\name{pastThinning}
\alias{pastThinning}
\title{Given a data frame and a pedigree, filter out crosses in the data frame that are already present in the pedigree}
\usage{
pastThinning(data, ped)
}
\arguments{
\item{data}{data frame with possible crosses to filter out}

\item{pedigree}{pedigree to not repeat}
}
\value{
Filtered data frame
}
\description{
Given a data frame and a pedigree, filter out crosses in the data frame that are already present in the pedigree
}
\examples{

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
