% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planCross.R
\name{planCross}
\alias{planCross}
\title{Create a mating plan from a group of parents}
\usage{
planCross(TargetPop, TargetPop2 = NULL, MateDesign = "half", Indiv2keep = NULL)
}
\arguments{
\item{TargetPop}{Individuals that should be mated.}

\item{TargetPop2}{Optional. Individuals that should be mated against the first list of individuals.}

\item{MateDesign}{indicates which type of mating design should be build up.
'full_p' all parents crossed with all parents, considering self of parents and reciprocal crosses.
'full' all parents crossed with all parents, considering reciprocal crosses.
'half_p' all parents crossed with all parents, considering self of parents but not reciprocal crosses.
'half' all parents crossed with all parents, with neither, self of parents or reciprocal crosses.}

\item{Indiv2keep}{Optional. Character vector with a list of candidate individuals. It will be
used to filter just those individuals to keep for built the mate plan.}
}
\value{
A data frame with all possible crosses.
}
\description{
A mating plan with all possible crosses for a given set of parents is generated.
}
\examples{
# 1.Loading the dataset.
data('datLines')

# 2.Using just a subset for time purposes
Parents = colnames(G)[1:15]

# 3.Creating the mating plan
plan1 = planCross(TargetPop = Parents,
                  MateDesign = 'half')

head(plan1,10)


# 4.Using two set of parents
Parents1 = colnames(G)[1:15]
Parents2 = colnames(G)[20:30]

# 5.Creating the mating plan
plan2 = planCross(TargetPop = Parents1,
                  TargetPop2 = Parents2,
                  MateDesign = 'half')

head(plan2,10)

}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
