% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanInbreedingRate.R
\name{RateInbreeding}
\alias{RateInbreeding}
\title{Calculates the inbreeding rates of a target population based on a SNP matrix.}
\usage{
RateInbreeding(W)
}
\arguments{
\item{W}{Marker matrix with the SNPs coded as 0,1,2.}
}
\value{
Inbreeding rate for a target population
}
\description{
The function implements the estimation for inbreeding rates based on the
proposition of Falconer and Mackay (1996). The amount of heterozygous is
measured as a proxy for the inbreeding rate for a target population.
}
