% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinePlan.R
\name{refinePlan}
\alias{refinePlan}
\title{Devolute mating plan to reduce relatedness}
\usage{
refinePlan(
  data,
  plan,
  K,
  n.cross = 50,
  max.cross = 4,
  min.cross = 2,
  culling.pairwise.k = 0.25,
  max.attempts = 100
)
}
\arguments{
\item{data}{data frame with possible crosses to filter out}

\item{plan}{current mating plan}

\item{K}{relationship matrix}

\item{n.cross}{number of crosses in the mating plan}

\item{max.cross}{maximum number of crosses per individual in the plan}

\item{min.cross}{minimum number of crosses per individual in the plan}

\item{culling.pairwise.k}{don't allow crosses with more than this threshold of relatedness. Default is 0.25.}

\item{max.attempts}{maximum number of attempts to improve the plan}
}
\value{
A list with a data frame with the attempts and a data frame with the improved mating plan
}
\description{
Devolute mating plan to reduce relatedness
}
\examples{

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
