% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relateThinning.R
\name{relateThinning}
\alias{relateThinning}
\title{Return a vector of individuals that maximizes criteria per cluster in the K matrix}
\usage{
relateThinning(K, Criterion, threshold = 0.5, max.per.cluster = 2)
}
\arguments{
\item{K}{relationship matrix. It could be a SNP-based or pedigree-based relationship matrix.}

\item{Criterion}{data frame with variable to maximize, i.e., estimated breeding
value, BLUP, genetic value, etc., for each candidate parent.}

\item{threshold}{K threshold to cluster.}

\item{max.per.cluster}{maximum number of entries per cluster to keep.}
}
\value{
A vector with genotype names
}
\description{
The function filter the individuals and cut off those with high relationship with each other.
It does that based on the relationship matrix, the criteria used to rank the individuals, a threshold for
group formation and a maximum number of individuals per cluster.
}
\examples{
  
# 1.Loading the data
data('datGeneric')

# 2.Criterion
Crit = data.frame(Id = colnames(G),
                  Criterion = Criterion[,1])

# 3.Thinning
parents2keep = relateThinning(K = G, 
                              Criterion = Crit, 
                              threshold = 0.5, 
                              max.per.cluster = 2)

# 4.List with the parents to keep
parents2keep

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
