% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relMatIBD.R
\name{relMatIBD}
\alias{relMatIBD}
\title{Compute Genomic Relationship Matrix using IBD Methods}
\usage{
relMatIBD(Markers, ploidy, sep = "_")
}
\arguments{
\item{Markers}{A numeric matrix of genotype data where rows represent markers
and columns represent individuals/haplotypes.}

\item{ploidy}{An integer indicating the ploidy level of the organism.}

\item{sep}{Character separator used in column names for AM method (default "_").}
}
\description{
Internal function to compute genomic relationship matrices
using Allele Matching (AM) method.
}
\keyword{internal}
