% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usef2Crosses.R
\name{usef2Crosses}
\alias{usef2Crosses}
\title{Generates the input for optimization algorithm}
\usage{
usef2Crosses(Usefulness, K)
}
\arguments{
\item{Usefulness}{data frame with the output of the usefulness functions (getusefA, getusefA_mt, getusefAD, getusefAD_mt).}

\item{K}{relationship matrix between all genotypes}
}
\value{
data frame with four columns to be used in the optimization function.
}
\description{
This is a useful function that takes the output of usefulness function and the information on relationship among individuals (G matrix) and
generates the input to the optimization function. The output is a data frame with four columns, Parent1, Parent2, Y, and K, ready to be the input of the function selectCrosses.
}
\examples{
\dontrun{
# 1.Loading the data
data(lines_geno)
data(lines_addEffects)
data(lines_GenMap)

# 2. Crossing plan
CrossPlan <- planCross(TargetPop = rownames(lines_geno))

# 3. Calculating the usefulness of trait number 1
usef_add <- getUsefA(MatePlan = CrossPlan,
                     Markers = lines_geno,
                     addEff = lines_addEffects[, 1],
                     Map.In = lines_GenMap,
                     propSel = 0.05,
                     Type = "RIL")


# 4. Creating relationship matrix
ScaleMarkers <- scale(lines_geno, scale = FALSE)

relMat <- (ScaleMarkers \%*\% t(ScaleMarkers)) / ncol(ScaleMarkers)

# 5. Input data
df4optimization <- usef2Crosses(Usefulness = usef_add, K = relMat)

# 6. Data table
head(df4optimization[[1]], 15) # For usefulness

head(df4optimization[[2]], 15) # For the mean
}

}
