% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G_ibd}
\alias{G_ibd}
\title{Compute the IBD Genomic Relationship Matrix}
\usage{
G_ibd(Markers, ploidy)
}
\arguments{
\item{Markers}{A numeric matrix of genotype data where rows represent markers
and columns represent individuals. Values should range from 0 to ploidy.}

\item{ploidy}{Integer. it indicates the ploidy level.}
}
\value{
A symmetric numeric matrix of dimensions n x n (where n is the number
of individuals) containing the genomic relationship coefficients.
}
\description{
Computes a IBD relationship matrix from IBD data.
}
\keyword{internal}
