% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pedigree.R
\docType{data}
\name{generic_pedigree}
\alias{generic_pedigree}
\title{generic_pedigree}
\format{
A data frame with 120 rows and 3 variables.
\describe{
\item{\code{id}}{Individuals' id}
\item{\code{mother}}{mother' id}
\item{\code{father}}{father' id}
}
}
\usage{
data(generic_pedigree)
}
\description{
A pedigree information for 100 individuals. The three columns represents the individuals, mother, and father identification, respectively.
}
\keyword{datasets}
