% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndex.R
\name{getIndex}
\alias{getIndex}
\title{Estimates a selection index based on the genotypes values and weights}
\usage{
getIndex(Criterion, Weights = NULL, Scale = TRUE)
}
\arguments{
\item{Criterion}{matrix containing the 'BLUP', 'ebv' or 'tgv' for more than one trait}

\item{Weights}{Row vector containing the weights for each trait}

\item{Scale}{logical. Default value is TRUE. If FALSE, it will not scale the traits}
}
\value{
A index based on the traits and weights.
}
\description{
Estimates a selection index based on the genotypes values and weights
}
\examples{

# 1. Loading the data
data('datGeneric')

# 2. Index
trait_index = getIndex(Criterion=Criterion, 
                       Weights= c(0.5,0.5), 
                       Scale = TRUE)
# 3. Output
head(trait_index, 10)

}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
