% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCrossMPA.R
\name{getMPA}
\alias{getMPA}
\title{Estimates the expected mean parental average for a set of crosses.}
\usage{
getMPA(MatePlan, Criterion, K = NULL, Weights = NULL, Scale = TRUE)
}
\arguments{
\item{MatePlan}{vector of possible crosses to make.}

\item{Criterion}{data frame with the parents id and the BLUPs of the individuals.}

\item{K}{relationship matrix between all individuals.}

\item{Weights}{vector with the weights for each trait. Only used when more than one trait is given.}

\item{Scale}{Boolean. If TRUE, the traits values will be scaled. Default is TRUE. Only used when more than one trait is given.}
}
\value{
A data frame with all possible crosses from the MatePlan, their mean parental average, and covariance from the relationship matrix.
}
\description{
The expected mean parental average (MPA) is estimated between each pair of individuals presented in the MatePlan. It uses the Criterion argument input for the estimation. Then, the Criterion should
be BLUPs for the individuals presented in the MatePlan. For more than one trait, the MPA is estimated by an index, a linear combination among the traits. Weights should be informed for each trait in this case. The relationship matrix
(K) here presented as an argument is for creating the output in the form of the input for the optimization function (selectCrosses).
}
\examples{
\dontrun{
# 1.Loading the data
data(lines_Geno)
data(lines_IndBLUP)

# 2.Criterion
Crit <- data.frame(Id = lines_IndBLUP[, 1],
                   Criterion = lines_IndBLUP[, 2])

# 3. Creating relationship matrix
relMat <- (lines_Geno \%*\% t(lines_Geno)) / ncol(lines_Geno)

# 4.Mating Plan
CrossPlan <- planCross(TargetPop = lines_IndBLUP[, 1])

# 5.Single trait mean parental average
ST_mpa <- getMPA(MatePlan = CrossPlan,
                 Criterion = Crit,
                 K = relMat)

head(ST_mpa, 20)

# 6. Criterion for MTM
CritMT <- data.frame(Id = lines_IndBLUP[, 1],
                     Criterion = lines_IndBLUP[, 2:3])


# 7. Multi trait mean parental average
MT_mpa <- getMPA(MatePlan = CrossPlan,
                 Criterion = CritMT,
                 K = relMat,
                 Scale = TRUE,
                 Weights = c(0.3, 0.7))

head(MT_mpa, 20)
}

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com} & Marco A. Peixoto, \email{marco.peixotom@gmail.com}
}
