% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsefA_mt.R
\name{getUsefA_mt}
\alias{getUsefA_mt}
\title{Prediction of usefulness for a set of crosses (Multi additive trait)}
\usage{
getUsefA_mt(
  MatePlan = NULL,
  Markers = NULL,
  addEff = NULL,
  Map.In = NULL,
  propSel = 0.05,
  Type = NULL,
  Generation = 0,
  Weights = NULL
)
}
\arguments{
\item{MatePlan}{data frame with the two columns indicating the crosses to predict.}

\item{Markers}{matrix with markers information for all candidate parents,
coded as 0,1,2.}

\item{addEff}{data frame with additive marker effects for each trait.}

\item{Map.In}{data.frame with the mapping information, i.e., chromosome, positioning and SNP id.}

\item{propSel}{Value representing the proportion of the selected individuals.}

\item{Type}{which kind of system of mating: "DH": doubled haploids lines or
"RIL": Recombinant inbred lines.}

\item{Generation}{integer. Indicates the generation where the parents come from. According to Lehermeier et al. (2017)
DH from F1 generation is '0' and RILs from F1 generation is '1'.}

\item{Weights}{Row vector containing the weights for each trait}
}
\value{
A data frame with means, variances, and usefulness for each pair of
crosses presented in the MatePlan.
}
\description{
Predicts usefulness component for a set of crosses. It accounts for additive-based traits.
}
\examples{
# 1.Loading the dataset.
data('datLines')

# 2.Using just a subset for time purposes
Parents = colnames(G)[1:15]

# 3.Creating the mating plan
plan = planCross(TargetPop = Parents,
                 MateDesign = 'half')

# 4.Calculating the usefulness for trait number 1
usef_add = getUsefA_mt(MatePlan = plan,
                       Markers = Markers,
                       addEff = addEff,
                       Map.In = Map.In,
                       propSel = 0.05,
                       Type = 'DH',
                       Weights = c(0.4,0.6))

head(usef_add, 10)

}
\references{
\emph{Lehermeier, C., de los Campos, G., Teyssèdre, S., & Schön, C. C. (2017). Genetic gain increases by applying the usefulness criterion with improved variance prediction in selection of crosses. Genetics, 207(4), 1651-1661.}

\emph{Bonk, S., Reichelt, M., Teuscher, F., Segelke, D., & Reinsch, N. (2016). Mendelian sampling covariability of marker effects and genetic values. Genetics Selection Evolution, 48(1), 1-11.}
}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
