% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planCross.R
\name{planCross}
\alias{planCross}
\title{Create a mating plan from a group of parents}
\usage{
planCross(TargetPop, MateDesign = "half", TargetPop2 = NULL, Indiv2keep = NULL)
}
\arguments{
\item{TargetPop}{Individuals that should be mated.}

\item{MateDesign}{indicates which type of mating design should be build up.
'full_p': this option creates a crossing plan where all parents are crossed with all parents, considering self of parents and reciprocal crosses.
'full': this option creates a crossing plan where all parents are crossed with all parents, considering reciprocal crosses.
'half_p': this option creates a crossing plan where all parents are crossed with all parents, considering self of parents but not reciprocal crosses.
'half': this option creates a crossing plan where all parents are crossed with all parents, with neither, self of parents or reciprocal crosses.
'maxAvoid': this option creates a maximum avoidance of inbreeding crossing plan (Kimura & Crow, 1963).
'circularPlan': this option creates a circular crossing plan (Kimura & Crow, 1963).}

\item{TargetPop2}{Optional. Individuals that should be mated against the first list of individuals.}

\item{Indiv2keep}{Optional. Character vector with a list of candidate individuals. It will be
used to filter just those individuals to keep for built the mate plan.}
}
\value{
A data frame with all possible crosses.
}
\description{
A mating plan with all possible crosses for a given set of parents is generated. When two target populations are used (TargetPop and TargetPop2) the North Carolina 2 mating design (Comstock and Robinson, 1952) is implemented.
}
\examples{
\dontrun{
# 1.Loading the dataset.
Parents <- rep(paste0("UF", rep(1:20)))

# 2.Creating the mating plan
plan1 <- planCross(TargetPop = Parents,
                   MateDesign = "half")


head(plan1, 10)

# 3.Using two set of parents
Parents1 <- Parents[1:10]
Parents2 <- Parents[11:20]

# 4.Creating the mating plan
plan2 <- planCross(TargetPop = Parents1,
                   TargetPop2 = Parents2,
                   MateDesign = "half")


head(plan2, 10)
}


}
\references{
\emph{Peixoto, Amadeu, Bhering, Ferrao, Munoz, & Resende Jr. (2024). SimpleMating:  R-package for prediction and optimization of breeding crosses using genomic selection. The Plant Genome, e20533.https://doi.org/10.1002/tpg2.20533}
}
\author{
Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
