% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relateThinning.R
\name{relateThinning}
\alias{relateThinning}
\title{Return a vector of individuals that maximizes criteria per cluster in the K matrix}
\usage{
relateThinning(K, criterion, threshold = 0.5, max.per.cluster = 2)
}
\arguments{
\item{K}{relationship matrix}

\item{criterion}{data frame with variable to maximize}

\item{threshold}{K threshold to cluster}

\item{max.per.cluster}{maximum number of entries per cluster to keep}
}
\value{
A vector with genotype names
}
\description{
Return a vector of individuals that maximizes criteria per cluster in the K matrix
}
\examples{

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
