% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCrosses.R
\name{selectCrosses}
\alias{selectCrosses}
\title{Select a mating plan that maximizes the criteria given some parameters}
\usage{
selectCrosses(
  data = NULL,
  n.cross = 200,
  max.cross = 4,
  min.cross = 2,
  max.cross.to.search = 1e+05,
  K = K,
  culling.pairwise.k = NULL
)
}
\arguments{
\item{data}{data frame with possible crosses to filter out.}

\item{n.cross}{number of crosses in the mating plan.}

\item{max.cross}{maximum number of crosses per individual in the plan.}

\item{min.cross}{minimum number of crosses per individual in the plan (this is the target, some parents might be used just one time).}

\item{max.cross.to.search}{maximum number of rows in the input to search for the solution.}

\item{K}{relationship matrix.}

\item{culling.pairwise.k}{don't allow crosses with more than this threshold of relatedness. Default is NULL, don't use it.}
}
\value{
A data frame with all possible crosses.
}
\description{
Select a mating plan that maximizes the criteria given some parameters
}
\examples{
# 1.Loading the data
data('datLines')

# 2.Mating Plan
CrossPlan = planCross(TargetPop = colnames(G) )

# 3. Criterion
Crit = data.frame(Id = colnames(G),
                  Crit = Criterion[,1])

# 4.Single trait mean parental average
ST_mpa = getMPA(MatePlan = CrossPlan, 
                Criterion = Crit, 
                K=G)

# 5 Crosses selected
maxGainPlan = selectCrosses(data = ST_mpa,
                            K = G,
                            n.cross = 80,
                            max.cross = 3,
                            min.cross = 1,
                            culling.pairwise.k = 1) 

# 6. Crosses parameters
maxGainPlan[[1]]

# 7. Mating Plan
maxGainPlan[[2]]


}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
