% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCrosses.R
\name{selectCrosses}
\alias{selectCrosses}
\title{Select a mating plan that maximizes the criteria given some parameters}
\usage{
selectCrosses(
  data = NULL,
  n.cross = 200,
  max.cross = 4,
  min.cross = 2,
  max.cross.to.search = 1e+05,
  K = K,
  culling.pairwise.k = NULL
)
}
\arguments{
\item{data}{data frame with possible crosses to filter out}

\item{max.cross}{maximum number of crosses per individual in the plan}

\item{min.cross}{minimum number of crosses per individual in the plan (this is the target, some parents might be used just one time)}

\item{max.cross.to.search}{maximum number of rows in the input to search for the solution}

\item{K}{relationship matrix}

\item{culling.pairwise.k}{don't allow crosses with more than this threshold of relatedness. Default is NULL, don't use it.}

\item{n.crosses}{number of crosses in the mating plan}
}
\value{
A data frame with all possible crosses.
}
\description{
Select a mating plan that maximizes the criteria given some parameters
}
\examples{

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com}
}
