% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCrosses.R
\name{setCrosses}
\alias{setCrosses}
\title{Creates a mating plan that maximizes the mean parental average of a set of crosses}
\usage{
setCrosses(
  Criterion,
  MateDesign = "half",
  n.cross = 200,
  max.cross = 4,
  min.cross = 2,
  max.cross.to.search = 1e+05,
  Weights = NULL,
  Scale = TRUE
)
}
\arguments{
\item{Criterion}{data frame with the parents and the criterion used for each parent.}

\item{MateDesign}{indicates which type of mating design should be build up.
'full_p' all parents crossed with all parents, considering self of parents and reciprocal crosses.
'full' all parents crossed with all parents, considering reciprocal crosses.
'half_p' all parents crossed with all parents, considering self of parents but not reciprocal crosses.
'half' all parents crossed with all parents, with neither, self of parents or reciprocal crosses.}

\item{n.cross}{number of crosses in the mating plan.}

\item{max.cross}{maximum number of crosses per individual in the plan.}

\item{min.cross}{minimum number of crosses per individual in the plan (this is the target, some parents might be used just one time).}

\item{max.cross.to.search}{maximum number of rows in the input to search for the solution.}

\item{Weights}{vector. Weights for each trait for building the selection index to be used.}

\item{Scale}{logical. If 'TRUE' it will scale the traits. Otherwise, will not scale the traits beforehand. The default is TRUE.}
}
\value{
A data frame with all possible crosses.
}
\description{
For a set of individuals (parents), this function creates a mating plan with the mean parental average of the individuals. The best option here is to use BLUPs from each candidate to be parent.
Multiple trait can also be used, and weights should be given. No constrains to inbreeding is used. If there is information regarding relationship among individuals, please, use the function selectCrosses.
}
\examples{
\dontrun{
# 1.Loading the data
data(lines_IndBLUP) # BLUPs/Criteria

# 2. Criterion for one trait
Crit <- data.frame(Id = lines_IndBLUP[, 1],
                   Crit = lines_IndBLUP[, 2] )


# 3. Cross
Matplan <- setCrosses(Criterion = Crit,
                      MateDesign = "half",
                      n.cross = 100,
                      max.cross = 10,
                      min.cross = 1)


# 4. Stats
Matplan[[1]]

# 5.Mating plan
Matplan[[2]]


# 6. Criterion for two traits
CritMT <- data.frame(Id = lines_IndBLUP[, 1],
                     Crit = lines_IndBLUP[, 2:3])


# 7. Cross
MatplanMT <- setCrosses(Criterion = CritMT,
                        MateDesign = "half",
                        n.cross = 100,
                        max.cross = 10,
                        min.cross = 1,
                        Weights = c(0.5, 0.4))


# 8. Stats
MatplanMT[[1]]

# 9.Mating plan
MatplanMT[[2]]
}

}
\author{
Rodrigo R Amadeu, \email{rramadeu@gmail.com} Marco Antonio Peixoto, \email{marco.peixotom@gmail.com}
}
