% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsefA.R, R/getUsefAD.R, R/getUsefAD_mt.R,
%   R/getUsefA_mt.R
\name{theta}
\alias{theta}
\title{\code{theta}
Function to calculate the recombination matrix from a genetic map based on Haldane (1909).}
\usage{
theta(map)

theta(map)

theta(map)

theta(map)
}
\arguments{
\item{map}{data frame with three columns: chromosome number, chromosome position, and marker number}
}
\description{
\code{theta}
Function to calculate the recombination matrix from a genetic map based on Haldane (1909).

\code{theta}
Function to calculate the recombination matrix from a genetic map based on Haldane (1909).

\code{theta}
Function to calculate the recombination matrix from a genetic map based on Haldane (1909).

\code{theta}
Function to calculate the recombination matrix from a genetic map based on Haldane (1909).
}
