% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usef2Crosses.R
\name{Usef2Crosses}
\alias{Usef2Crosses}
\title{Generates the input for optimization}
\usage{
Usef2Crosses(Usefulness = NULL, K = NULL)
}
\arguments{
\item{Usefulness}{data frame with the output of the usefulness functions (getusefA, getusefA_mt, getusefAD, getusefAD_mt).}

\item{K}{relationship matrix between all genotypes}
}
\value{
data frame with four columns to be used in the optimization function.
}
\description{
With the output of usefulness function, and the information on relationship,
it generates the input to the optimization function
}
\examples{
# 1.Loading the dataset.
data('datLines')

# 2.Using just a subset for time purposes
Parents = colnames(G)[1:15]

# 3.Creating the mating plan
plan = planCross(TargetPop = Parents,
                 MateDesign = 'half')

# 4.Calculating the usefulness for trait number 1
usef_add = getUsefA(MatePlan = plan,
                    Markers = Markers,
                    addEff = addEff[,1],
                    Map.In = Map.In,
                    propSel = 0.05,
                    Type = 'DH')

head(usef_add, 10)

# 5. Main table
MainTab = Usef2Crosses(Usefulness=usef_add, K=G)

head(MainTab)

}
